Remove subfunction definition from cmdline shell with exec
Calling exec in a function that defines a sub function with local
variables is not supported for Python<=2.7.6. When called the
following exception will be thrown:

	SyntaxError: unqualified exec is not allowed in function
	'shell' it contains a nested function with free variables

However, this was used in the `verdi shell` command implementation.
To maintain backwards compatibility with older python version, that
are often still installed by default on OS'es or clusters,  the sub
function is refactored to be defined as a top level function.
