+++ docs/requirements_for_rtd.txt
@@ -50,7 +50,7 @@ psutil==5.4.5
 pymatgen==2018.4.20
 pyparsing==2.2.0
 pytest-cov==2.5.1
-pytest==3.5.1
+pytest==3.6.0
 python-dateutil==2.7.2
 python-memcached==1.59
 python-mimeparse==1.6.0
